+++ Control/Joint/Effects/Either.hs
@@ -30,8 +30,13 @@ instance Transformer (Either e) where
 instance Applicative u => Liftable (Either e) (UT (Either e) u) where
 	lift = UT . pure
 
+instance Liftable (Either e) t => Liftable (Either e) (UT t u) where
+	lift = lift
+
 instance Liftable (Either e) u => Liftable (Either e) (TU t u) where
 	lift = lift
 
 instance Liftable (Either e) u => Liftable (Either e) (TUT t u t') where
 	lift = lift
+
+type Failable e t = Liftable (Either e)