+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 * Define `Liftable` class to lift many effects in schemes and other effects
 * Create `Abilities` and `Effects` umbrella modules
 * Rename `Configured` datatype back to `Reader`
+* Define `Failable`, `Optional`, `Configured` and `Stateful` constraints